Move all benchmorks into a single file
